Move old files to deprecated folders
