+++ deprecated/gpt-3.5-replication.py
@@ -15,7 +15,7 @@ EXP_ID = 2
 RUN_DOC_EMBEDDINGS = False
 
 DATA_PATH = './data/'
-CHROMA_PATH = "chroma"
+CHROMA_PATH = "./chroma/chroma_text-embedding-ada-002-500"
 PROMPT_TEMPLATE = """
 You are a helpful assistant assessing the quality of academic papers. Answer the question with YES or NO. Write nothing else before or after.
 Answer the question based only on the following context:
