+++ deprecated/groq_new.py
@@ -27,7 +27,7 @@ EXP_ID = 2
 RUN_DOC_EMBEDDINGS = False
 
 DATA_PATH = './data/'
-CHROMA_PATH = './chroma_' + EMBEDDING_MODEL + '-' + str(CHUNK_SIZE)
+CHROMA_PATH = './chroma/chroma_' + EMBEDDING_MODEL + '-' + str(CHUNK_SIZE)
 
 PROMPT_TEMPLATE = """
 You are a helpful assistant assessing the quality of academic papers. Answer the question with YES or NO. Write nothing else before or after.
