+++ src/old_data_out/750-gpt-3.5-turbo-text-embedding-3-large-p3.csv
