+++ src/old_data_out/CoT-250-gpt-3.5-turbo-text-embedding-3-large-p3.csv
