+++ src/old_data_out/CoT-gpt-3.5-turbo-text-embedding-3-large-p2.csv
