+++ src/old_data_out/CoT-gpt-4o-text-embedding-3-large-p1.csv
