+++ src/old_data_out/gemma2-9b-it-text-embedding-3-large-p2.csv
