+++ src/old_data_out/gemma2-9b-it-text-embedding-ada-002-p3.csv
