+++ src/old_data_out/gemma2_original/gemma2-9b-it-text-embedding-ada-002-p2-original.csv
