+++ src/old_data_out/gpt-3.5-turbo-text-embedding-3-large-p1.csv
