+++ src/old_data_out/gpt-3.5-turbo-text-embedding-ada-002-p2.csv
