+++ src/old_data_out/gpt-4o-text-embedding-3-large-p3.csv
