+++ src/old_data_out/gpt-4o-text-embedding-ada-002-p3.csv
